requirements.txt
google-generativeai
python-dotenv
docling
python-pptx
Pillow
numpy
pandas
python-dateutil
rapidfuzz
tqdm